+++ libraries/base/System/Posix/Types.hs
@@ -71,6 +71,30 @@ module System.Posix.Types (
 #if defined(HTYPE_RLIM_T)
   CRLim(..),
 #endif
+#if defined(HTYPE_BLKSIZE_T)
+  CBlkSize(..),
+#endif
+#if defined(HTYPE_BLKCNT_T)
+  CBlkCnt(..),
+#endif
+#if defined(HTYPE_CLOCKID_T)
+  CClockId(..),
+#endif
+#if defined(HTYPE_FSBLKCNT_T)
+  CFsBlkCnt(..),
+#endif
+#if defined(HTYPE_FSFILCNT_T)
+  CFsFilCnt(..),
+#endif
+#if defined(HTYPE_ID_T)
+  CId(..),
+#endif
+#if defined(HTYPE_KEY_T)
+  CKey(..),
+#endif
+#if defined(HTYPE_TIMER_T)
+  CTimer(..),
+#endif
 
   Fd(..),
 
@@ -108,6 +132,9 @@ import GHC.Base
 import GHC.Enum
 import GHC.Num
 import GHC.Real
+#if defined(HTYPE_TIMER_T)
+import GHC.Float
+#endif
 -- import GHC.Prim
 import GHC.Read
 import GHC.Show
@@ -157,8 +184,38 @@ INTEGRAL_TYPE(CTcflag,HTYPE_TCFLAG_T)
 INTEGRAL_TYPE(CRLim,HTYPE_RLIM_T)
 #endif
 
--- ToDo: blksize_t, clockid_t, blkcnt_t, fsblkcnt_t, fsfilcnt_t, id_t, key_t
--- suseconds_t, timer_t, useconds_t
+#if defined(HTYPE_BLKSIZE_T)
+-- | @since 4.10.0.0
+INTEGRAL_TYPE_WITH_CTYPE(CBlkSize,blksize_t,HTYPE_BLKSIZE_T)
+#endif
+#if defined(HTYPE_BLKCNT_T)
+-- | @since 4.10.0.0
+INTEGRAL_TYPE_WITH_CTYPE(CBlkCnt,blkcnt_t,HTYPE_BLKCNT_T)
+#endif
+#if defined(HTYPE_CLOCKID_T)
+-- | @since 4.10.0.0
+INTEGRAL_TYPE_WITH_CTYPE(CClockId,clockid_t,HTYPE_CLOCKID_T)
+#endif
+#if defined(HTYPE_FSBLKCNT_T)
+-- | @since 4.10.0.0
+INTEGRAL_TYPE_WITH_CTYPE(CFsBlkCnt,fsblkcnt_t,HTYPE_FSBLKCNT_T)
+#endif
+#if defined(HTYPE_FSFILCNT_T)
+-- | @since 4.10.0.0
+INTEGRAL_TYPE_WITH_CTYPE(CFsFilCnt,fsfilcnt_t,HTYPE_FSFILCNT_T)
+#endif
+#if defined(HTYPE_ID_T)
+-- | @since 4.10.0.0
+INTEGRAL_TYPE_WITH_CTYPE(CId,id_t,HTYPE_ID_T)
+#endif
+#if defined(HTYPE_KEY_T)
+-- | @since 4.10.0.0
+INTEGRAL_TYPE_WITH_CTYPE(CKey,key_t,HTYPE_KEY_T)
+#endif
+#if defined(HTYPE_TIMER_T)
+-- | @since 4.10.0.0
+FLOATING_TYPE_WITH_CTYPE(CTimer,timer_t,HTYPE_TIMER_T)
+#endif
 
 -- Make an Fd type rather than using CInt everywhere
 INTEGRAL_TYPE(Fd,CInt)
@@ -184,4 +241,3 @@ type ProcessID      = CPid
 type FileOffset     = COff
 type ProcessGroupID = CPid
 type Limit          = CLong
-