+++ libraries/base/include/CTypes.h
@@ -29,11 +29,17 @@ newtype T = T B deriving (ARITHMETIC_CLASSES, INTEGRAL_CLASSES) \
                 deriving newtype (Read, Show);
 
 #define INTEGRAL_TYPE_WITH_CTYPE(T,THE_CTYPE,B) \
-newtype {-# CTYPE "THE_CTYPE" #-} T = T B deriving (ARITHMETIC_CLASSES, INTEGRAL_CLASSES) \
-                                          deriving newtype (Read, Show);
+newtype {-# CTYPE "THE_CTYPE" #-} T = T B \
+    deriving (ARITHMETIC_CLASSES, INTEGRAL_CLASSES) \
+    deriving newtype (Read, Show);
 
 #define FLOATING_TYPE(T,B) \
 newtype T = T B deriving (ARITHMETIC_CLASSES, FLOATING_CLASSES) \
                 deriving newtype (Read, Show);
+
+#define FLOATING_TYPE_WITH_CTYPE(T,THE_CTYPE,B) \
+newtype {-# CTYPE "THE_CTYPE" #-} T = T B \
+    deriving (ARITHMETIC_CLASSES, FLOATING_CLASSES) \
+    deriving newtype (Read, Show);
 
 #endif