+++ libraries/base/changelog.md
@@ -14,6 +14,8 @@
   * `New modules `Data.Bifoldable` and `Data.Bitraversable` (previously defined
     in the `bifunctors` package) (#10448)
 
+  * `Data.Either` now provides `fromLeft` and `fromRight` (#12402)
+
 ## 4.9.0.0  *May 2016*
 
   * Bundled with GHC 8.0