+++ libraries/base/changelog.md
@@ -26,6 +26,8 @@
   * Add `Semigroup` instance for `IO`, as well as for `Event` and `Lifetime`
     from `GHC.Event` (#12464)
 
+  * Add `Data` instance for `Const` (#12438)
+
 ## 4.9.0.0  *May 2016*
 
   * Bundled with GHC 8.0