+++ libraries/base/changelog.md
@@ -30,6 +30,8 @@
 
   * Added `Eq1`, `Ord1`, `Read1` and `Show1` instances for `NonEmpty`.
 
+  * Raw buffer operations in `GHC.IO.FD` are now strict in the buffer, offset, and length operations (#9696)
+
 ## 4.9.0.0  *May 2016*
 
   * Bundled with GHC 8.0